Client

Description

The Client module

Documentation

client :: QSem -> QSem -> IORef [[(Int, Int)]] -> IORef [(Int, Int)] -> IORef Bool -> IORef Bool -> [IORef [(Int, Int)]] -> IORef Bool -> IO ()

The client that implements the clientside gamelogic

clientloop :: Int -> QSem -> QSem -> String -> IORef [(Int, Int)] -> Socket -> IORef Bool -> [IORef [(Int, Int)]] -> IORef Bool -> IO ()

Implements the logic of the main game

fire :: QSem -> QSem -> IORef [(Int, Int)] -> Socket -> IORef [(Int, Int)] -> IORef [(Int, Int)] -> IORef [(Int, Int)] -> IORef Bool -> IO ()

Is called when the player is allowed to fire

wait :: Socket -> IORef [(Int, Int)] -> IORef [(Int, Int)] -> IORef [(Int, Int)] -> IORef Bool -> IO ()

Is called when the player is waiting for the enemeys choice

fieldsetting :: QSem -> QSem -> IORef [[(Int, Int)]] -> IORef [(Int, Int)] -> IORef Bool -> IO ()

Is called at the beginning of the game

forbiddentuples :: [[(Int, Int)]] -> Set (Int, Int)

Creates a Set of forbidden fields

memberInSet :: [(Int, Int)] -> Set (Int, Int) -> Bool

Checks if the players choice is a member of the Set of forbidden fields

check' :: [(Int, Int)] -> [[(Int, Int)]] -> Bool

Checks if the maximal number of fields is valid